+++ UnityTimelineBuilder/Packages/com.hidano.unity-timeline-builder/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+
+- Samples~/BuildMenu を追加しました (TimelineBuilder.Build() をメニューから実行するサンプル)。
+
 ## [0.1.0] - 2026-08-20
 
 - パッケージの初期構成と UPM マニフェストを追加しました。